Move payloads into payloads.ts
